+++ dropbox-hacking-photo-manager-client/src/components/editableTextField.tsx
