+++ dropbox-hacking-photo-manager-client/src/components/navigate.tsx
