+++ dropbox-hacking-photo-manager-client/src/components/photoTile.tsx
