+++ dropbox-hacking-photo-manager-client/src/pages/legacy/days/calendar/index.tsx
@@ -4,7 +4,7 @@ import { useEffect } from "react";
 import { useCountsByDate } from "@/context/countsByDateContext";
 import logRender from "@/logRender";
 import SamePageLink from "@/components/samePageLink";
-import Navigate from "@/pages/legacy/days/navigate";
+import Navigate from "@/components/navigate";
 import MonthBox from "./monthBox";
 import TwelveMonths from "./twelveMonths";
